Safari-609.1.11.1/LayoutTests/http/tests/security/mixedContent/insecure-form-in-iframe.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that has a form with an insecure action.  We should
*not* get a mixed content callback because the main frame is HTTP and the form
doesn't contaminate the child iframe's security origin with mixed content.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-form.html"></iframe>
</body>
</html>
